create storage service interface and implement local storage service so that it can be moved to minio later easily. Also develop mapping stage

+++ src/main/java/com/xjtlu/bio/service/SampleService.java
@@ -126,7 +126,7 @@ public class SampleService {
         }
     }
 
-    public void testInsertDuplicate(String sampleName, int projectId) {
+    public void testInsertDuplicate(String sampleName, long projectId) {
         BioSample bioSample = new BioSample();
         bioSample.setSampleName(sampleName);
         bioSample.setProjectId(projectId);
